+++ pkg/structs/disk.go
@@ -4,7 +4,7 @@ package structs
 type Save struct {
 	UGI      string `json:"ugi" validate:"required" label:"ugi"`
 	Token    string `json:"token" validate:"required" label:"token"`
-	SaveSlot int    `json:"save_slot" validate:"required,min=1,max=10" label:"save_slot"`
+	SaveSlot uint8  `json:"save_slot" validate:"required,min=1,max=10" label:"save_slot"`
 	SaveData any    `json:"save_data" validate:"required,max=10000" label:"save_data"`
 }
 
@@ -12,5 +12,5 @@ type Save struct {
 type Load struct {
 	UGI      string `json:"ugi" validate:"required" label:"ugi"`
 	Token    string `json:"token" validate:"required" label:"token"`
-	SaveSlot int    `json:"save_slot" validate:"required,min=1,max=10" label:"save_slot"`
+	SaveSlot uint8  `json:"save_slot" validate:"required,min=1,max=10" label:"save_slot"`
 }
